Added quick return checks in some WIndow functions.
Window object now remembers the default state that is first passed in during creation, and will return to it when coming out of fullscreen windowed. This fixes the issue of the window coming out of fullscreen and being a window larger than the screen.

Made the fullscreen switch remember to reapply the title and icon when switching to windowed.